rename courselistsidebar -> inspector
