--- src/layouts/CourseListSideBar/CourseListSideBar.tsx
@@ -1,5 +1,0 @@
-export const DarsSideBar: React.FC = () => {
-  return <div>Course Details</div>;
-};
-
-export default DarsSideBar;
