--- src/layouts/CourseListSideBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CourseListSideBar';
