+++ src/layouts/Inspector/CourseDetails.tsx
