+++ src/layouts/Inspector/Inspector.tsx
@@ -1,0 +1,5 @@
+export const Inspector: React.FC = () => {
+  return <div>Course Details</div>;
+};
+
+export default Inspector;
